bpf: introduce atomic_add()
The GCC BPF support does not provide a way to emit an atomic add
instruction (BPF xadd) without the fetch-flag.  But since older
kernels do not support xadd with the fetch flag, the generated
BPF programs are rejected by the BPF verifier.

The atomic_add() macro emits a raw xadd instruction that works on
all kernels.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>