provider: support (opaque) private data for providers
Commit 2ba7c3bd ("Add a CPC provider") added an opaque private data
pointer to the provider implementation struct.  In retrospect, that
does not seem to be the best place because the implementation struct
is really meant for callback function pointers.  The provider struct
itself is a better place for this.  SDT providers will also make use
of the ability to store private data for providers.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>