pid, uprobes: improve horrendously inefficient uprobe scanning
The uprobe scanning loop was meant to be an efficient loop that scanned
the set of uprobes only once.  Unfortunately, because it is invoked
repeatedly to create each pid's uprobes in turn, it ends up parsing
all the probes N times if you invoke dtrace with N distinct processes.
Since each such parse involves multiple rounds of memory allocation
and string manipulation to decode encoded probe names etc, this is
a definite waste of time. Worse yet, since the uprobe list can change
at any time, there is not even any guarantee that we're parsing the
same thing each time!

Fix this by having dt_pid_create_usdt_probes() call another function,
dt_pid_scan_uprobes(), to pull in the list of known DTrace uprobes; it
does this only once and stores the parsed probespecs in a new list of
dt_pid_probespecs in the dtp.  The probespecs still change a bit on each
call to dt_pid_create_usdt_probes, because it stuffs a pid in there, but
everything derived from uprobe_events is parsed only once and (in this
implementation, currently) never changes thereafter, no matter how many
probes are created.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>