test: replace gawk-specific function with more generic function
The match(s, t, arr) form of the match function is specific to GAWK
and therefore is best avoided in favour of the more generic match()
function combined with tokenization.

The script was also not functioning correctly because the left
paranthesis of the action block wasn't on the same line as the
pattern it was supposed to be associated with.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>