dtprobed: support DOF > 64KiB
dtprobed's ioctl handler is a state machine that reads in client DOF in
multiple phases:

 DTP_IOCTL_START: read the struct dof_helper_t

 DTP_IOCTL_HDR: read the struct dof_hdr_t, giving us the size of the DOF

 DTP_IOCTL_DOFHDR: read the DOF itself

 DTP_IOCTL_DOF: (everything is read; actually parse the DOF)

These various phases are triggered by issuing a fuse_reply_ioctl_retry()
to ask FUSE to read things out of the ioctl() caller for us and call us
back with the new stuff.  We can read from any address in the client (we
are root), but that doesn't mean we can read any *length* -- the kernel
refuses to read beyond a certain length and alas that length is
customizable via mount options (or, for CUSE, via kernel command-line
options) and we can't easily tell what it is.

I think it reasonable to say that anyone who *reduces* the length below
the default is asking for it: I can't find any CUSE or FUSE-using code
anywhere on the Internet that does anything more than hardwire this
value to 128KiB. For paranoia, we use 64KiB ourselves -- that's still a
lot of probes in one go.

We add a new state, DTP_IOCTL_DOFCHUNK, and read in one piece of DOF
per chunk until all the chunks are read in, concatenating them before
handing it off to the parser.

A new test checks this stuff: it has 206KiB of DOF which is enough to
test "first chunk", "neither first nor last chunk" and "last chunk"
cases: the DOF itself isn't formatted such that this equates to "first
probe, probe in the middle, last probe" but if any of those chunks is
missing or misaligned the string tables will be fubared and all the
probe names will be wrong, triggering a test failure (at best).

Orabug: 35411920
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>