link: do not modify the input object files unless changed
process_obj() and the various implementations of dt_modtext() go to
considerable lengths to not modify the input .o files unless they
actually need to change (dtrace -G differs from a linker in this,
since it modifies its input object files as well as creating an
output).

But if it finds probes, it unconditionally declares the file modified
and (eventually) calls elf_update() on it, even if all the modifications
were already done by a prior invocation of dtrace -G on the same inputs.
This leads to needless re-make(1)s and is entirely unnecessary.  (The
output is still modified regardless, of course.)

(nca: wrote test, commit log)

Orabug: 35417184
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>