map_value_or_null: add check for NULL pointer in load_var()
When we look up a dynamic pointer (e.g., associative array or thread-local
variable), the BPF verifier sees that we get back a "map_value_or_null".
Ultimately, we will check the validity of the value.  If we first add an
offset to it, however, the BPF verifier objects.  This is a little strange,
since the verifier is fine adding an integer to 0 or to a map value.
Nevertheless, that's the situation.

Check the value.  If 0, assign 0.  This seems like it doesn't do anything,
but it allows the BPF verifier to branch and handle map_value_or_null as
two separate cases.

Add tests.  While this fix addresses dynamic variables (associative arrays
and thread-local variables), add tests for global and local variables also.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>