dtprobed: prevent inter-request contamination of state
On fairly rare occasions, one sometimes sees dtprobed saying things like
this:

46157: dtprobed: helper size incorrect: expected at least 80, not 0

or

25815: dtprobed: DOF too small: expected at least 64, not 0

The message varies, but the size is always 0.  Whenever this happens
one object's-worth of incoming DOF is thrown out.

The cause is, it turns out, simple.  CUSE ioctls are handled via a
multistage request/response protocol, where the CUSE server (dtprobed)
repeatedly calls fuse_reply_ioctl_retry() and returns and is immediately
called back with the result of extracting a bit more info from the
client's address space: this process is fast but still involves a
protocol roundtrip to the FUSE device, so it's not instantaneous.  To
carry state across this call FUSE has a userdata pointer that is shared
across calls to all the FUSE callbacks.

Now *obviously* CUSE isn't going to interleave the results of this
request/response process with further requests coming in from new
clients, right? I mean even the FUSE examples are written assuming that
this won't happen.

It does, and this problem is, while not terribly common, common enough
that I'm not the only one who's seen it. You can be halfway through a
request/response cycle and then some new PID can come blundering in and
you have to handle that as well, in parallel.

So we can't just pass a single userdata state vector around and have it
track everything because all our requests are serialized, since they're
*not* serialized.  So switch to managing them by hand: maintain a
dt_list_t of manually-malloced userdatas, reverse-insertion-sorted by
pid so that unless the pid wraps the pid we want is almost always at the
head of the list.  We need to clean up this list periodically because
processes can be killed in the middle of their ioctl() so we can get
stale stuff in the list: to save a bit of time and make sure the cleanup
code doesn't rust, *only* free in a periodic cleanup cycle: the cycle
goes through the list and frees every element with no associated process
or whose state indicates that it's not in the middle of a
request/response cycle right now.

For now, clean up every 128 requests -- rare enough that it's
amortized-free, common enough that even when pids wrap we hardly have to
traverse any stale ones to get to the one we care about.

Add a test for the cleanup code (and a bit of a load-test for dtprobed)
that just kicks off a bunch of processes with probes.  This won't fail
itself but if things go wrong it might provoke a dtprobed crash and a
failure of later usdt tests, which is not a valueless thing (since the
cleanup code is otherwise not tested at all, and it found a bug when I
wrote it, so let's keep it).

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>