test: Have manyprobes clean up uprobes
Once dtprobed cleans uprobes up automatically, this hack will no longer
be needed.  Meanwhile, we have to clean up lest a few runs of this test
lead to denial of service.

Note that the test did try to clean up uprobes but it referred to
uprobe_events with an incorrect pathname (leading to ineffectual cleanup
and, on some systems, test failures) and it cleaned up all uprobes, even
ones not created by this test.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>