test: rework main lockstat test
The main lockstat test (tst.lockstat.sh) didn't need to be implemented
using a shell script.  It has been reworked to be a regular .d script
with a .t trigger executable script.  Because of varying kernel options
between aarch64 and x86_64, each arch gets its own expected output.

Test failures may be reported if kernel configurations differ from those
used for UEK6 and UEK7.  More specific tests (configured based on the
kernel configuration of the runtime kernel) will need to be developed to
make this type of testing more effective.

The test has been renamed as tst.lockstat-summary.d to reflect that it is
a test the summarizes the overall implementation of lockstat probes.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>