test: Loosen tolerance for switchrate test
The timing expected for the switchrate test is rather approximate.
We want a "short" switchrate to run in a "short" amount of time and
a "long" one to run "long."  That's about it.

The actual times observed tend to be a little higher than the expected
times.  On some systems (ARM OL9?), the discrepancy is slightly higher.
This test, however, is simply concerned with whether the switchrate
option is working.

Loosen the tolerance for the test.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>