spec, dtprobed: restart on install/update if already running
If we don't restart dtprobed when a yum update happens, we're going to
keep running an older version, with all its bugs and/or limitations.  I
thought %systemd_postun handled this, but in fact it doesn't: you need
to use the barely-documented %systemd_postun_with_restart instead.  Of
course, that doesn't really work: being in %postun, the *package being
uninstalled* already needs to have it.  So do a try-restart in %post
instead, so that the new version gets it. If we do a try-restart before
the enable/start, it should work fine even on first install, when it's
not already running.

While we're at it, fix %systemd_postun to not mention .targets,
only .services: it doesn't make sense to stop targets :)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>