ctf: add BTF-to-CTF convertor
For kernels that do not provide CTF data, kernel type information can be
obtained at runtime as BTF data.  We pull in that data and generate
equivalent CTF data from it.

Note that BTF is not a true replacement for CTF data.  I.e. it does not
currently provide type information for kernel variables.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>