dtprobed: add the DOF stash
The DOF stash is a runtime directory (/run/dtrace by default) in which
the dtprobed daemon will keep track of incoming DOF.  There's a comment
describing it at the top of dtprobed/dof_stash.c, but in brief we have three
directories under this:

- stash/dof/, which contains raw DOF named by mapping
- stash/dof-pid/, which tracks DOF on a PID-by-PID and mapping-by-mapping
  basis. This contains subdirectories $pid/$mapping containing both the same
  raw DOF (hardlinked) and parsed DOF split up by probe, in
  $pid/$mapping/parsed/$prv:$mod:$fun$:prb.
- probes/, which contains parsed DOF for each individual probe in
  /run/dtrace/probes/$pid/$prv$pid/$mod/$fun/$prb to aid consumer globbing
  of probes; in future we may introduce further symlinks not in per-pid
  directories to ease matching of probes across processes.

Consumers can use inotify watches to detect newly added or removed DOF or
probes: because all the DOF is in one directory, a very small number of
watches is needed to keep track of all the DOF in the system, and only one
watch per PID is needed to track it on a PID-by-PID basis. (Currently,
nothing is doing this, but it is now a possibility where it never was
before.)

In a future commit, dtprobed will emit all the DOF it receives into this
structure, removing it when processes exit, and DTrace will use it to create
and remove probes.

dof_stash.c itself provides functions to add new raw DOF to the stash, to
add a stream of parsed DOF corresponding to a given piece of raw DOF to the
stash, to remove DOF as processes issue DTRACEHIOC_REMOVE requests, to check
for any DOF that needs reparsing to handle changes to struct dof_parsed, and
to remove entries from the DOF stash that is left behind by processes that
died uncleanly without issuing DTRACEHIOC_REMOVEs. (Currently, we don't deal
with newly-started processes that reuse the same PID as such a process, or
processes that leave DOF behind on exec() and then add more: this
improvement will come soon.)

It does not do any DOF parsing: that is left up to the caller (dtprobed.c).

All the writes to the dof_stash are done via the *at() functions to avoid
any risk of bugs causing the file creations and most especially deletions it
carries out from leaking outside /run/dtrace.  (We are not concerned about
symlink attacks etc because only root can write to /run/dtrace or any
directories underneath it.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>