dtprobed: empty the DOF stash for exec()ed processes
The DOF stash, and USDT in general, currently have problems with fork()ed
and exec()ed processes. The worst of these problems is with exec().  Since
exec() replaces a process image with another atomically, ELF destructors do
not get to run, so dtprobed never receives a DTRACEHIOC_REMOVE request for
any of the DOF in it.  If the exec()ed process also contains DOF, this DOF
will add to the DOF already in the stash, and DTrace will think that the
exec()ed process contains all the probes in the original process in addition
to its own.

The fix is fairly simple, though it involves a bit of annoying refactoring:
record the (dev, ino) of the executable mapping of every process
contributing DOF in a new dof-pid file "exec-mapping", then compare this
with the executable mapping of the contributing procerss when new DOF is
contributed.  Any exec() will necessarily change this and we can clean out
the old DOF dir (hence the refactoring, since this adds a second thing
beyond stale process cleanup that wants to delete whole PIDs out of the DOF
stash).

Testing this is a little painful because DTrace proper does not detect
exec() and rescan for new USDT probes (it would also need to wait until the
new process had run its drti ELF constructor); but it can be done.

Technically there is still one case in which the DOF can go stale: if a
process does dlopen()s, then exec()s itself.   This is fixable (by detecting
new-DOF contributions from a mapping that we already know about for this
PID), but is a slightly distinct fix with a different testcase, so is best
put in a separate commit.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>