test: Increase period in cpc allcpus test
Some kernels have problems with clock firings due to throttling.
The CPC allcpus test uses a similar software event because that
event is so ubiquitous.  But, while we have migrated many tests
away from clock ticks due to their fickleness, this test inherently
depends on this unfortunately unstable event.

Nevertheless, it seems that the problems with this event are
mitigated when events fire less often, and we only need one
firing per CPU for this test.

Therefore, increase the period.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>