cg: ensure argument NULL pointer check detects constant 0 values
The dt_cg_arg_to_tstring() function was generating code for a runtime
NULL pointer check if the passed argument was a pointer or a string.
But if the passed argument was a NULL constant, that check was never
generated because the argument was an integer in that case.

The new code determines the need for the NULL pointer check based on
the datatype of the argument (as specified in the function prototype)
instead.

Tests are included based on inet_ntoa6().

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>