Replace dt_variable_read()
The dt_variable_read() function is only used in a few places, all of
them involving reading data from a data record.  Callers adjust the
base address using the record offset, and then pass in the data size
from the record as well.  The new dt_read_scalar() function takes a
base address and a pointer to a record descriptor which is more in
line with the nature of data record descriptors.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>