Dedicate space for call stacks
In the dctx->mem memory block, storage for a call stack and storage
for tstrings overlapped.  The rationale was that a stack() action and
temporary strings would never co-exist.

However, the call stack storage wasn't being used for stack() (or ustack
or jstack) actions at all.  It was being used for stackdepth (and
ustackdepth and jstackdepth) built-in variables.  And they could
co-exist with temporary strings.

So make the call stack storage its own area within dctx->mem.

Add a test to check for the overlap bug.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>