proc: rip out waitfd() and hit waitpidding thread with a signal instead
For a long time we have used waitfd() to allow our process-monitoring
threads to wait for messages from DTrace proper (via a pipe) and state
changes from the monitored process (via Pwait(), which calls waitpid())
without having to engage in CPU-chewing polling loops and without inducing
latency in the monitored process or in responses to DTrace. (Both matter:
the proxying mechanism is quite fine-grained. The main DTrace thread can
mess around in libproc as much as it likes, with individual Ptrace() calls
being automatically proxied to the monitoring thread down the pipe. So
single operations from the perspective of DTrace's main thread can involve
dozens of proxy calls. Excessive latency would be bad.)

But waitfd() is ugly enough that it's never going to go upstream: not only
was it already rejected in 2010, but it needs to add really ugly hacks to
the waitqueue mechanism and even disable some assetions because it bloats
some core polling data structures beyond a cacheline (theoretically slowing
them down, though I've never been able to measure anything).

So it would be very nice to get rid of it.

The solution is an old Unix horror: EINTR. Everyone hates it, everyone wraps
long-running syscalls in EINTR loops to evade it... and here it does just
what we want. We find an unused realtime signal (the only WIP: actually hunt
one down rather than just stamping on one) and unmask it in the monitoring
thread, then drop the whole waitfd thing entirely and typically wait in
waitpid() (only waiting blocked on poll() on the proxy pipe if we're
explicitly not listening to the process right now). When we send a proxy
message to a monitoring thread, we hit it with this signal, which causes
waitpid() to exit with EINTR -- and once it does that we can check the proxy
pipe and process any messages, with no polling loops or added latencies.

We use SIGRTMIN as the signal by default: if the caller is one of those few
that actually uses realtime signals for something, we provide a new
dtrace_set_internal_signal API function that the caller can invoke before
calling dtrace_open to reset the signal to some other value that the caller
is not using (specified as a number to be added to SIGRTMIN, since we *do*
require that it's a realtime signal).

This requires an additional hook, analogous to the existing ptrace_lock_hook
(which is used to take out the dpr_lock around Ptrace() calls, so that no
other Ptrace() calls can happen for that process at the same time). This
one, the waitpid_lock_look, is used to *drop* the lock around the call to
waitpid(), because the waitpid() call may now take a long time, and the
proxy calling mechanism has to take out the dpr_lock (because it protects
the dpr_msg_cv condvar that mediates the proxy call). But we can't drop the
dpr_lock around the call to Pwait() as a whole because that call also
invokes all the breakpoint handlers, and *that* work requires the dpr_lock
to have been already taken out by Pwait()s caller. So we need another
hook. Thankfully the two hooks are never nested!

This does have one race we haven't fixed yet, because the fix is quite
involved: see the next commit.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>